Basic one-step flow is working.
The client generates an attestation key, sends it and the endorsement
key to the server, and gets back a sealed certificate and a wrapped key.

The client can unwrap the key with the TPM endorsement encryption key,
and then unseal the certificate.

The CA identifies machines using their endorsement public key.  After
thinking it through, there's not really a better alternative --- it
seems like TPM support for endorsement certificates and platform
certificates is spotty.  Even if it wasn't, the fact that the TPM or
machine was manufactured by a certain party isn't a good basis for
allowing it membership in your cluster.